// mobile/src/api/generated/schemas/trainingPlanResponse.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * GrowIt.API | v1
 * OpenAPI spec version: 1.0.0
 */
import type { PlanDaySummaryResponse } from './planDaySummaryResponse';

export interface TrainingPlanResponse {
  id: string;
  name: string;
  /** @nullable */
  notes: string | null;
  isActive: boolean;
  days: PlanDaySummaryResponse[];
}